Support shared drives
Google drive does not allow service accounts to
have their own quota, so files must be added to
a shared drive on a paid workspace account. Files
will count against the workspace quota.

Added security: specify the shared drive by ID in
the environment. We previously were potentially
vulnerable to someone discovering the service
account address and sharing folders with the
correct name to the service. Now we limit the
folder search to the shared drive.

We might want to create separate shared drives
for dev, staging, and prod to manage quota use
separately.