Add an office address API
- pull the phonebook down during build time
- office addresses are keyed by lastname and state

We will need to look up senators using the info
we get from 5calls, and some senators are
listed in that database using nicknames (ie. "Chuck"
vs. "Charles").

If we don't care about photos of the senators,
we could use this database for everything.

Why build time: because I honestly don't expect
the gov server to reliably up or fast, and so
don't want it in the serving path.

added playwrite tests, but they should not run
in the continuous integration. Restrict vitest
to the test/ directory to keep it from trying to
run the e2e tests.