use salutation in the letter if we have it
If we can get the official last name from the phone
book, use it in the letter salutation instead of the
full name from 5calls. If we don't have it fall back
to full name.

also rework address and reps in a passable envelope
object to clean up some type warnings and prep for
signable objects.