add envelope validation
I don't see a reason to worry about replay attacks,
but I don't want clients to rewrite envelope information,
such as a senator's names.

I'm not validating the date. The drafts are dropped
into weekly folders chosen by server time, so at
worst someone changes the date on a letter but that
won't change when it is seen or delivered? Maybe
reconsider this?