add a SHOW_CAPTCHA flag
USE_CAPTCHA controls server-side validation. This
should be on on production and off in development
unless you are specifically debugging validation.

SHOW_CAPTCHA controls the client-side presentation of
the widget. This should be on in both production and
development unless the widget is slowing down
development cycles with challenges.

If reCAPTCHA is causing a production outage, set
SHOW_CAPTCHA to false in production to disable both
rendering and validation.